requirements.txt
streamlit==1.25.0
tensorflow==2.10.0
numpy==1.24.2
Pillow==9.5.0
matplotlib==3.7.0
